+++ tests/__init__.py
@@ -1,5 +1,6 @@
-from .character_tests import CharacterTests
-from .role_tests import RoleTests
-from .weapon_tests import WeaponTests
-from .server_tests import ServerTests
-from .faction_tests import FactionTests
+# from .character_tests import CharacterTests
+# from .role_tests import RoleTests
+# from .weapon_tests import WeaponTests
+# from .server_tests import ServerTests
+# from .faction_tests import FactionTests
+from .base_tests import BaseTestCase, MyTestCase
