+++ tests/character_tests.py
@@ -5,15 +5,6 @@ from nocapapi.models import Character, RosterUser, Server, Role, Faction, Weapon
 from nocapapi.views.character import CharacterSerializer
 
 class CharacterTests(APITestCase):
-    # Add any fixtures you want to run to build the test database
-    fixtures = ['users', 'tokens', 'rosterusers',  'weapons', 'factions', 'roles', 'servers', 'characters',]
-    
-    def setUp(self):
-        # Grab the first rosteruser object from the database and add their token to the headers
-        self.rosteruser = RosterUser.objects.first()
-        token = Token.objects.get(user=self.rosteruser.user)
-        self.client.credentials(HTTP_AUTHORIZATION=f"Token {token.key}")
-
     def test_create_character(self):
         """Create character test"""
         url = "/characters"
