+++ tests/role_tests.py
@@ -5,14 +5,6 @@ from nocapapi.models import  RosterUser, Role
 from nocapapi.views.role import RoleSerializer
 
 class RoleTests(APITestCase):
-    # Add any fixtures you want to run to build the test database
-    fixtures = ['users', 'tokens', 'rosterusers', 'roles' ]
-    
-    def setUp(self):
-        # Grab the first rosteruser object from the database and add their token to the headers
-        self.rosteruser = RosterUser.objects.first()
-        token = Token.objects.get(user=self.rosteruser.user)
-        self.client.credentials(HTTP_AUTHORIZATION=f"Token {token.key}")
     def test_get_roles(self):
         """Get roles test"""
         url = "/roles"
